[feedback] FlightProgressBar 함수 분리
